Basic Character UI and Character Movement
Clicking on a Basic Character now brings up a basic version of the Character UI. From here, clicking on the move button allows the player to move the character according to the mouse position. The offsets for the UI are fixed and do not currently take into account screen resolution.